refactor

+++ src/getInteger.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-static int checkErr(const char *endptr, const char *buf, const long number) {
+static int checkErr(const char *endptr, const char *buf, const long number)
+{
     if (errno == ERANGE)
     {
         (void) fprintf(stderr, "\33[31m the number is too small or too large\33[m");
@@ -15,7 +16,7 @@ static int checkErr(const char *endptr, const char *buf, const long number) {
     else if (*endptr && *endptr != '\n')
     {
         (void) fprintf(stderr, "\33[31m wrong symbols were entered. "
-               "Fractional number or odd space after the number may have been entered\33[m");
+                               "Fractional number or odd space after the number may have been entered\33[m");
         return 0;
     }
     else if (errno != 0 && number == 0)
@@ -26,11 +27,11 @@ static int checkErr(const char *endptr, const char *buf, const long number) {
     return 1;
 }
 
-long getInteger(const char *outputString)
+int getInteger(const char *outputString)
 {
     char buf[1024];
     int successFlag;
-    long number;
+    int number;
     do
     {
         printf("%s", outputString);
@@ -44,5 +45,5 @@ long getInteger(const char *outputString)
         number = strtol(buf, &endptr, 10);
         successFlag = checkErr(endptr, buf, number);
     } while (!successFlag);
-    return number;
+    return (int) number;
 }
